test/integration/gles/replay: Move check functions into own file
